+++ docs/CHANGELOG.md
@@ -5,7 +5,14 @@
 * Added **widgets** support (experimental): 
   * a websocket opened from the front-end that communicates
     to the kernel, and through it to the users cells.
-  * API to use it in `gonb/gonbui/widgets`. 
+  * API to use it in `gonb/gonbui/widgets`.
+  * API to communicate with front-end in `gonb/gonbui/comms`.
+    * `Listen[T](address)` function added to create a channel listening
+      to front-end updates.
+  * API to manipulate the DOM in `gonb/gonbui/dom`. 
+    * `Persist()` added to persist transient changes to the DOM -- meaning
+      dynamically generated HTML and widgets will show up when exporting
+      to HTML or when running `nbconvert`.
 * Added "%wasm" support (experimental): 
   * Allows compiling cell to WASM and running that in the notebook. One
     can write widgets like this. **Experimental**: there are some use cases are not 100% clear. See